Added a different counter type
Previously everything was done on a clear counter, now added a container counter and refactored the code to make it much simpler, by handling the logic for general type counter in the BaseCounter which would be extended by each specific type of counters.

For the container counter, added animation the closes and opens the door after the player interacts with it and picks up the object.