Added the ability to place objects on a plate
Implemented two ways to place objects on a plate, either by holding the plate and interacting with a counter that has one of the objects that is allowed to be placed on a plate, or by holding one of those objects and interacting with a counter that has a plate on top.

Added clear visual to the added objects on the plate.